Add BopProvider selection and move Bop class to dedicated package
