+++ simdesk-web/src/main/java/de/sustineo/simdesk/entities/bop/Bop.java
@@ -1,4 +1,4 @@
-package de.sustineo.simdesk.entities;
+package de.sustineo.simdesk.entities.bop;
 
 import lombok.*;
