+++ simdesk-web/src/main/java/de/sustineo/simdesk/views/generators/InactiveBopPartNameGenerator.java
@@ -1,7 +1,7 @@
 package de.sustineo.simdesk.views.generators;
 
 import com.vaadin.flow.function.SerializableFunction;
-import de.sustineo.simdesk.entities.Bop;
+import de.sustineo.simdesk.entities.bop.Bop;
 
 public class InactiveBopPartNameGenerator implements SerializableFunction<Bop, String> {
     @Override
